requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pymongo==4.6.0
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & Requests
httpx==0.25.2
aiofiles==23.2.1

# QR Code generation
qrcode[pil]==7.4.2

# Image processing (for AI validation)
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Caching & Rate Limiting
redis==5.0.1
slowapi==0.1.9

# Environment
python-dotenv==1.0.0

# Additional dependencies
pyjwt==2.8.0
cryptography==41.0.7
faker==20.1.0
